requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.7
django-cleanup==8.1.0
django-excel==0.0.10
djangorestframework==3.15.2
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==22.0.0
lml==0.1.0
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
texttable==1.7.0
tzdata==2024.1
whitenoise==6.7.0
geographiclib==2.0
geopy==2.4.1